+++ minishell/prueba.txt
@@ -1,0 +1,3 @@
+hola$
+buenas$
+sisi
